receipts program
a system to help speed up the process of creating receipts for your customers.
In this project, we will be storing the names and prices of a furniture store’s catalog in variables. You will then process the total price and item list of customers, printing them to the output terminal.